requirements.txt
anthropic==0.5.0
openai-whisper==20231117
python-dotenv==1.0.0
numpy<2
google-generativeai==0.3.1
openai==1.3.0